bot/pyproject.toml
[project]
name = "mt5-trading-bot"
version = "1.0.0"
description = "Production-ready MT5 automated trading bot with logging and monitoring"
requires-python = ">=3.9"

dependencies = [
    "MetaTrader5>=5.0.45",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14b",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "Flask>=2.3.0",
    "flask-cors>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
